Improved launch framework, there is now one pathway
Before the simulate and main paths were totally seperate.
Now they are the same with one argument for simulating passed
down throughout. It is relatively intelligent with what should
launch in each case. It is possible that #TODO some of the packages
being used in the launch files aren't in the appropriate package.xml
files.