Got recording to work.
To get recording working, I had to seriously decrease the quality of
everything. I think I am ok with decreasing the quality of the fisheye
camera. I also think I am ok with decreasing the quality of the RGB
recording to match that of the depth. But IDK. I also reduced frame
rates, which I am not OK with, but will have to live with for now?